requirements.txt
spotipy==2.24.0
pandas==2.2.3
numpy==2.1.3
sqlalchemy==2.0.36
plotly==5.24.1
kaleido==0.2.1
python-dotenv==1.0.1
pytest==8.3.3